backend-services/auth/requirements.txt
uvicorn==0.16.0
fastapi==0.71.0
pydantic==1.9.0
python-multipart==0.0.5
environs==9.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
databases[postgresql]==0.5.3
sqlalchemy==1.4.29
psycopg2-binary==2.9.3

  